Socket in own file
